requirements.txt
bcrypt==3.2.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
paramiko==2.8.0
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
regex==2021.11.10
requests==2.26.0
scp==0.14.1
six==1.16.0
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
